backend/requirements.txt
# KnowAI Backend Dependencies
# Core
FastAPI>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
pydantic-settings>=2.0

# Database & storage
sqlalchemy>=2.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Task queue
celery[redis]>=5.3.0
redis>=5.0.0

# Vector & search
qdrant-client>=1.7.0

# Storage
minio>=7.2.0

# HTTP client for LLM
httpx>=0.27.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0

# Parsers
PyMuPDF>=1.23.0
python-docx>=1.0.0

# Embeddings (optional; stub works without it)
# sentence-transformers>=2.2.0
